Refactor: Groups~ 클래스 Group~ 클래스로 통합
Groups~ 클래스 Group~ 클래스로 통합
