--- src/main/java/com/prography1/eruna/service/GroupsService.java
@@ -1,52 +1,0 @@
-package com.prography1.eruna.service;
-
-import com.prography1.eruna.domain.entity.GroupUser;
-import com.prography1.eruna.domain.entity.Groups;
-import com.prography1.eruna.domain.entity.User;
-import com.prography1.eruna.domain.repository.GroupUserRepository;
-import com.prography1.eruna.domain.repository.GroupsRepository;
-import com.prography1.eruna.domain.repository.UserRepository;
-import com.prography1.eruna.response.BaseException;
-import com.prography1.eruna.response.BaseResponseStatus;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class GroupsService {
-    private final UserRepository userRepository;
-    private final GroupsRepository groupsRepository;
-    private final GroupUserRepository groupUserRepository;
-
-    public Boolean isValidCode(String code){
-        return groupsRepository.existsByCode(code);
-    }
-    public Groups findByCode(String code){
-        return groupsRepository.findByCode(code).orElseThrow(()-> new BaseException(BaseResponseStatus.INVALID_GROUP_CODE));
-    }
-
-    public boolean isDuplicatedNickname(String code, String nickname) {
-        Groups group = findByCode(code);
-        return groupUserRepository.existsByGroupsAndNickname(group, nickname);
-    }
-
-    public GroupUser joinGroupUser(String code, String uuid, String nickname, String phoneNum){
-        Groups group = findByCode(code);
-        User user = userRepository.findByUuid(uuid).orElseThrow(() -> new BaseException(BaseResponseStatus.USER_NOT_FOUND));
-        GroupUser.GroupUserId groupUserId = GroupUser.GroupUserId.builder()
-                .groupId(group.getId())
-                .userId(user.getId())
-                .build();
-        GroupUser groupUser = GroupUser.builder()
-                .groupUserId(groupUserId)
-                .user(user)
-                .groups(group)
-                .nickname(nickname)
-                .phoneNum(phoneNum)
-                .build();
-        return groupUserRepository.save(groupUser);
-    }
-
-}
