+++ src/main/java/com/prography1/eruna/web/GroupReqDto.java
@@ -26,7 +26,7 @@ public class GroupReqDto {
         @Schema(description = "전화번호", example = "01012345678")
         private String phoneNum;
 
-        @Schema(description = "알람 설정", example = "")
+        @Schema(description = "알람 설정", example = "알람 설정 스키마 참고")
         private AlarmInfo alarmInfo;
     }
 
@@ -49,5 +49,18 @@ public class GroupReqDto {
 
         @Schema(description = "벌칙", example = "커피 쏘기!")
         private String penalty;
+    }
+
+    @Schema(title = "유저 그룹 참가")
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class GroupJoinUserInfo {
+        @Schema(description = "그룹에 참여하는 유저 uuid", example = "1111-AAAA-UUID-1234")
+        private String uuid;
+        @Schema(description = "새 유저 닉네임", example = "닉네임1")
+        private String nickname;
+        @Schema(description = "유저 전화번호", example ="01012345678")
+        private String phoneNum;
     }
 }
