--- src/main/java/com/prography1/eruna/web/GroupsController.java
@@ -1,43 +1,0 @@
-package com.prography1.eruna.web;
-
-import com.prography1.eruna.response.BaseException;
-import com.prography1.eruna.response.BaseResponse;
-import com.prography1.eruna.response.BaseResponseStatus;
-import com.prography1.eruna.service.GroupsService;
-import com.prography1.eruna.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/group")
-@RequiredArgsConstructor
-public class GroupsController {
-    private final GroupsService groupsService;
-    private final UserService userService;
-
-    @PostMapping("/{code}")
-    public BaseResponse<String> userJoinGroup(@PathVariable String code, @RequestBody GroupsReqDto.GroupJoinUserInfo groupJoinUserInfo){
-        if(!groupsService.isValidCode(code)) throw new BaseException(BaseResponseStatus.INVALID_GROUP_CODE);
-        String uuid = groupJoinUserInfo.getUuid();
-        if(!userService.isUserExist(uuid)) throw new BaseException(BaseResponseStatus.INVALID_UUID_TOKEN);
-        String nickname = groupJoinUserInfo.getNickname();
-        if(groupsService.isDuplicatedNickname(code, nickname)) throw new BaseException(BaseResponseStatus.DUPLICATED_NICKNAME);
-
-        groupsService.joinGroupUser(code, uuid, nickname, groupJoinUserInfo.getPhoneNum());
-        return new BaseResponse<>("ok");
-    }
-
-
-    @ExceptionHandler(BaseException.class)
-    public BaseResponse<?> handleBaseException(BaseException e) {
-        System.err.println(e.getClass());
-        return new BaseResponse<>(e.getStatus());
-    }
-
-    @GetMapping("/{code}/nickname-valid/{nickname}")
-    public BaseResponse<GroupResDto.IsValidNickname> isValidNickname(@PathVariable String code, @PathVariable String nickname){
-        if(groupsService.isDuplicatedNickname(code, nickname)) return new BaseResponse<>(new GroupResDto.IsValidNickname(false));
-
-        return new BaseResponse<>(new GroupResDto.IsValidNickname(true));
-    }
-}
